Initial commit of EuiTableOfRecords component
`EuiTableOfRecords` is a high level component that aims to easy the way one creates a table of... (wait for it....) records!!!

What is a record? Think of a record in a data store - typically it represents an entity with a very clear schema. It is something that can be presented in a tabular form.

The goal of this high level component is to make the consumer not think about design. Instead, the consumer only need to define the functional requirements of the table (features), the data, and the type of interaction the user should have with the shown records.

This high level component is as stateless as it needs to be. Meaning, all the management of the data (where it's coming from), record selection and loading pages... all of these are expected to be managed externally to this component. Typically one would use a container component to wrap around this table that will either manage this state internally to that component, or use other state stores such as Redux.

This initial commit includes:

- A first implementation of the `EuiTableOfRecords`
- Support for pagination
- Support for record selection
- Support for custom rendering of the record data

This commit also comes with a dedicated documentation section in the EUI doc site.