[EuiTableOfRecords] core model change
Previously the config enabled you to define multiple separate callbacks for pagination and sorting. The problem with that approach is that it still left too much for the user to manage. Specifically, certain behaviour that we'd like to promote and control from the `EuiTableOfRecords` component, was left to the user to implement - For example, one of the design patterns decision we had is to reset pagination when sorting changes. Unfortunately this cannot be done from within the component with multiple separate callbacks.

For this reason, the config and model props changes quite a bit. Now there's a clear separation between the data itself and the criteria that describes and loaded the data. The new model structure looks like this:

```js
{
  data: {
    records: [], // list of records
    totalRecordCount: number
  },
  criteria?: {
    page?: {
      index: number,
      size: number
    },
    sort?: {
      key: string,
      direction: 'asc' | 'desc'
    }
  }
}
```

In addition, there now only a single callback that asks the consumer to update the data - `onModelCriteriaChange`. This enables the table component to decide how the criteria change based on the interaction of the user - in other words, the table now controls its own behaviour.

Nice side effects:

- Pagination is enabled automatically when the `page` criteria is provided in the model (no need to explicitly configure pagination in the "config")
- Sorting is enabled automatically when one of the columns is marked as `sortable` (no need to explicitly configure sorting` in the "config")
- Rendering a table that shows all data (without pagination) is a simple as passing in the data - that is `model.data` (no need to pass in additional single page meta data as before).
- No more need for the `Page` construct

Examples:

- added a proper intro to the examples page - explaining the `config` and the `model` concepts. Also providing high level explanation of what exactly is a "High Level Component" (as it's relatively new to EUI).
- Updated all the examples based on the new model