[TableOfRecords] Added search and filtering support
It is not possible to configure search for the table. This comes in the form of a search bar that sits on top of the table. The search bar displays a search box and optionally a set of filters.

Search is based on a simple query model. When enabled, the query will part of the model criteria:

```
model : {
  ...
  criteria: {
    ...
    search: {
       query: 'query goes here'
    }
  }
}
```
When the search query changes (based on user input), the new search query will be available as part of the model criteria received in `onDataCriteriaChange(criteria)` callback. The consumer can then use this query to load and provide new records to the table.

The filters that can be configured will be shown next to the search box. These filters serve as a visual aid to manipulate the search query - that is, when a filter is clicked, the query in the search box will change to reflect the filter. Also, when the user types a query directly int he search box, the filters states will update accordingly - it's a bidirectional binding between the two - the search box and the filters, and this binding is driving by a single query model that sits behind them.

The query defines a simple query language that supports the following features:

- `default` clauses - simple terms are not directly associated with any specific record field (it is up to the query execution to define where these terms will be searched for). For example, the following query indicates that the user would like to get all records with "john" but are without "doe" - `john -doe`.
- `field` clauses - terms that are associated with named record fields (that is, the terms will be search in the defined fields). For example, the following will find all records that have "john" in the "firstName" field and don't have "doe" in the "lastName" field - `firstName:john -lastName:doe`.
- `is` clauses - this indicate that the record should match a "flag". Typically these refer to boolean fields that are associated with the records (e.g. "open", "on", "online", etc...). For example, the following query will find all issues that are opened but not assigned - `is:open -is:assigned`.

Here's a query that looks up all production watches that were triggered and that associated with the terms "website" - `type:watch is:triggered tag:production website`

Currently there are two supported filter types:

- `field_value_selection` - this filter is associated with a `field` clause (that is, it manipulates a `field` clause in the query). It is therefore associated with a field and enables the user to choose values to match on this field.

- `is` - this filter (as the name slightly suggests) is associated with an `is` clause - it manipulates a `is` clause in the query. It is associated with a (boolean) field as the flag to check.